+++ components/atoms/Button/index.stories.js
@@ -3,8 +3,8 @@ import { Button } from './index';
 
 export default { title: 'Button' };
 
-export const withText = () => <Button>Hello Button</Button>;
-
-export const withEmoji = () => (
-  <Button><span role="img" aria-label="so cool">😀 😎 👍 💯</span></Button>
-);   
+export const SmallButton = () => <Button small>Small</Button>;
+export const MediumButton = () => <Button medium>Medium</Button>;
+export const LargeButton = () => <Button large>Large</Button>;
+export const BlueLargeButton = () => <Button large background="blue">Large</Button>;
+export const ColoredTextButton = () => <Button large background="lightblue" fontColor="black">Large</Button>;
